chore: Add unsaved changes indicator to form
This commit adds functionality to display an indicator for unsaved changes in the form. When a user modifies any input field in the form, the field is marked as "changed" and the form is given the class "unsaved-changes". This allows the user to easily identify which fields have been modified and indicates that there are unsaved changes in the form.